Homework "JSON Processing" modified. "CarDealer" Database Query and Export Data part finished(Task 6). This homerwork is ready. All tasks are fulfilled.

+++ Homeworks/08.JSON Processing/CarDealer/CarDealer.Models/Car.cs
@@ -22,7 +22,7 @@
 
         public string Model { get; set; }
 
-        public double? TravelledDistance { get; set; }
+        public float? TravelledDistance { get; set; }
 
         public virtual ICollection<Part> Parts
         {
